Final Security Hardening: Rounds 5 & 6 - Map limits, Sleep sanitization, ReDoS protection, and String overflow fixes.
- Implemented coordinate limits (+/- 100,000), Z-level limits (100), and chunk limits (10,000 per Z) in Map.cs.
- Sanitized sleep() duration with 24-hour clamp and NaN/Infinity handling.
- Added 1-second match timeout to Regex constructions in DreamObjectRegex.cs.
- Fixed integer overflow in string concatenation length check in DreamValue.cs.
- Added reproduction tests for new vulnerabilities in VulnerabilityTests.cs.
- Unified VM stack management under Push() with MaxStackSize (65536) limit.
- Hardened DreamList with MaxListSize (1,000,000) and pre-allocation.
- Implemented 64MB result limit for string operations.
- Added MaxLocalVariables (1024) and parent type bounds checks in CompiledJsonService.
- Verified all 138 tests pass.

+++ Engine/Shared/Models/DreamValue.cs
@@ -268,7 +268,7 @@ namespace Shared
                 var sA = a.ToString();
                 var sB = b.ToString();
 
-                if (sA.Length + sB.Length > 67108864)
+                if ((long)sA.Length + sB.Length > 67108864)
                     throw new System.InvalidOperationException("Maximum string length exceeded during concatenation");
 
                 return new DreamValue(sA + sB);
